Moved join requests into specific page
